backend/requirements.txt
# 7tỷ.vn FastAPI Backend Dependencies
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling
python-multipart==0.0.6
Pillow==10.1.0
aiofiles==23.2.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Excel export
openpyxl==3.1.2
xlsxwriter==3.1.9

# QR Code generation
qrcode[pil]==7.4.2

# Vietnamese text processing
unidecode==1.3.7

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.25

# Development
pytest==7.4.3
pytest-asyncio==0.21.1